[221227] jdbc3

+++ jdbc_3/src/com/mc/jdbc/member/dao/MemberDao.java
@@ -1,12 +1,16 @@
 package com.mc.jdbc.member.dao;
 
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.ArrayList;
+import java.util.List;
 
+import com.mc.jdbc.common.code.ErrorCode;
+import com.mc.jdbc.common.exception.DataAccessException;
+import com.mc.jdbc.common.exception.HandleableException;
 import com.mc.jdbc.common.util.JDBCTemplate;
 import com.mc.jdbc.member.dto.Member;
 
@@ -33,15 +37,7 @@ public class MemberDao {
 			try(ResultSet rset = stmt.executeQuery(query)){
 				
 				while (rset.next()) {
-					member = new Member();
-					member.setUserId(rset.getString("user_id"));
-					member.setPassword(rset.getString("password"));
-					member.setGrade(rset.getString("grade"));
-					member.setTell(rset.getString("tell"));
-					member.setEmail(rset.getString("email"));
-					member.setLeave(rset.getBoolean("is_leave"));
-					member.setRegDate(rset.getTimestamp("reg_date").toLocalDateTime());
-					member.setRentableDate(rset.getTimestamp("rentable_date").toLocalDateTime());
+					member = generateMember(rset);
 				}
 			}
 		} catch (SQLException e) {
@@ -51,7 +47,21 @@ public class MemberDao {
 		return member;
 	}
 
-	public int insertMember(Member member) {
+	private Member generateMember(ResultSet rset) throws SQLException {
+		Member member;
+		member = new Member();
+		member.setUserId(rset.getString("user_id"));
+		member.setPassword(rset.getString("password"));
+		member.setGrade(rset.getString("grade"));
+		member.setTell(rset.getString("tell"));
+		member.setEmail(rset.getString("email"));
+		member.setLeave(rset.getBoolean("is_leave"));
+		member.setRegDate(rset.getTimestamp("reg_date").toLocalDateTime());
+		member.setRentableDate(rset.getTimestamp("rentable_date").toLocalDateTime());
+		return member;
+	}
+
+	public void insertMember(Connection conn, Member member) {
 		
 		String sql = "insert into member(user_id, password, email, grade, tell, rentable_date)"
 				+ " values('"+ member.getUserId() +"',"
@@ -62,56 +72,46 @@ public class MemberDao {
 					+ " now() "	
 					+ ")";
 		
-		int res = 0;
-		
-		Connection conn = null;
 		Statement stmt = null;
 		
 		try {
-			conn = jdt.getConnection();
-			stmt = conn.createStatement();
-			res = stmt.executeUpdate(sql);
 			
-			jdt.commit(conn);
+			stmt = conn.createStatement();
+			int res = stmt.executeUpdate(sql);
+			
 		} catch (SQLException e) {
 			e.printStackTrace();
-			jdt.rollback(conn);
+			throw new DataAccessException();
 		} finally {
 			jdt.close(stmt);
-			jdt.close(conn);
 		}
-		
-		return res;
 	}
 	
 	//userId의 비밀번호를
 	//매개변수로 받아온 password로 수정하는 코드를 작성하시오. 
-	public int changePassword(String userId, String password) {
+	public int changePassword(Connection conn, String userId, String password) {
 		
-		Connection conn = null;
-		PreparedStatement pstm = null;
+		//' or 1=1 or user_Id = '
+		
+		Statement pstm = null;
 		int res = 0;
 
-		String sql = "update member set password = ? where user_id = ?" ;
-		
+		String sql = "update member set password = '"+ password +"' where user_id = '" + userId +"'";
+		System.out.println(sql);
 		try {
 			
-			conn = jdt.getConnection();
-			pstm = conn.prepareStatement(sql);
-			pstm.setString(1, password); 
-			pstm.setString(2, userId);
-			res = pstm.executeUpdate();
+			pstm = conn.createStatement();
+			res = pstm.executeUpdate(sql);
 			
-			jdt.commit(conn);
+			// runnable Exception은 예외처리가 강제되지 않기 때문에, catch를 만날 때 까지 계속 호출부로 전달
+			if(res > 1) throw new HandleableException(ErrorCode.OCCURS_ANORMALY_ABOUT_ROW_CNT);
 			
 		} catch (SQLException e) {
-			
 			e.printStackTrace();
-			jdt.rollback(conn);
+			throw new DataAccessException();
 			
 		} finally {
 			jdt.close(pstm);
-			jdt.close(conn);
 		}
 		
 		return res;
@@ -119,34 +119,57 @@ public class MemberDao {
 	
 	
 	// 원하는 사용자의 아이디를 삭제하는 메서드
-	public int deleteUser(String userId) {
+	public int deleteUser(Connection conn, String userId) {
 		
-		Connection conn = null;
-		PreparedStatement pstm = null;
+		//' or 1=1 or user_Id = '
+	
+		Statement pstm = null;
 		int res = 0;
-		String sql = "delete from member where user_id = ?" ;
+		String sql = "delete from member where user_id = '" + userId + "'" ;
 		
 		try {
 			
-			conn = jdt.getConnection();
-			pstm = conn.prepareStatement(sql);
-			pstm.setString(1, userId);
-			res = pstm.executeUpdate();
+			System.out.println(sql);
 			
-			jdt.commit(conn);
+			pstm = conn.createStatement();
+			res = pstm.executeUpdate(sql);
+			
+			if(res > 1) {
+				System.out.println(res);
+				throw new HandleableException(ErrorCode.OCCURS_ANORMALY_ABOUT_ROW_CNT);
+			}
 			
 		}  catch (SQLException e) {
 			
 			e.printStackTrace();
-			jdt.rollback(conn);
+			throw new DataAccessException();
 			
 		} finally {
 			jdt.close(pstm);
-			jdt.close(conn);
 		}
 		
 		return res;
 		
+	}
+
+	public List<Member> selectAllMember(Connection conn) {
+		
+		List<Member> members = new ArrayList<Member>();
+		
+		try {
+			
+			PreparedStatement pstm = conn.prepareStatement("select * from member");
+			ResultSet rset = pstm.executeQuery();
+			
+			while(rset.next()) {
+				members.add(generateMember(rset));
+			}
+			
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+		
+		return members;
 	}
